requirements.txt
# Core PDF processing
PyPDF2>=3.0.0
reportlab>=4.0.0

# Office document support
python-docx>=0.8.11
openpyxl>=3.1.0

# Data processing
pandas>=2.0.0

# PDF text extraction
pdfplumber>=0.9.0

# Image processing and OCR (optional)
Pillow>=10.0.0
pytesseract>=0.3.10

# Printing support (optional)
pycups>=2.0.0

# Development dependencies (uncomment for development)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# flake8>=6.0.0
# black>=23.0.0